+++ 1.20.4/NeoForge/src/main/java/fuzs/universalenchants/neoforge/client/UniversalEnchantsNeoForgeClient.java
@@ -1,8 +1,10 @@
 package fuzs.universalenchants.neoforge.client;
 
 import fuzs.puzzleslib.api.client.core.v1.ClientModConstructor;
+import fuzs.puzzleslib.neoforge.api.data.v2.core.DataProviderHelper;
 import fuzs.universalenchants.UniversalEnchants;
 import fuzs.universalenchants.client.UniversalEnchantsClient;
+import fuzs.universalenchants.data.client.ModLanguageProvider;
 import net.neoforged.api.distmarker.Dist;
 import net.neoforged.bus.api.SubscribeEvent;
 import net.neoforged.fml.common.Mod;
@@ -14,5 +16,8 @@ public class UniversalEnchantsNeoForgeClient {
     @SubscribeEvent
     public static void onConstructMod(final FMLConstructModEvent evt) {
         ClientModConstructor.construct(UniversalEnchants.MOD_ID, UniversalEnchantsClient::new);
+        DataProviderHelper.registerDataProviders(UniversalEnchants.MOD_ID,
+                ModLanguageProvider::new
+        );
     }
 }